Fixed loading message for advanced; fixed sidebar appearance; adjusted routing; enhanced size warning popups for all loaders

+++ web/backend/src/main/java/de/exbio/reposcapeweb/communication/reponses/WebGraphService.java
@@ -1192,7 +1192,7 @@ public class WebGraphService {
         if (user.get() != null)
             userGraph.remove(user.get());
         File thumb = historyController.getThumbnailPath(gid);
-        if (thumb.exists())
+        if (thumb!=null &&thumb.exists())
             thumb.delete();
         historyController.remove(gid);
     }
